requirements.txt
streamlit==1.33.0
pandas
numpy
matplotlib
streamlit-authenticator
toml
pathlib
mysql-connector-python
pyrebase
